Fix a concurrency issue
The text of the error `attempt to index local 'tree'`.  It happens when
you run a test for the first time, but might be not reprodecued, because
of nature of concurrency issues.

`StateTracker:update_running` runs sometimes before
`StateTracker:update_positions`. Therefore the field `positions` is nil.

The issue is prevented with a semaphore.